+++ grid2op/PlotGrid/PlotMatplot.py
@@ -7,7 +7,6 @@
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
 import io
-import copy
 import numpy as np
 
 import matplotlib.pyplot as plt
@@ -18,19 +17,7 @@ from grid2op.PlotGrid.PlotUtil import PlotUtil as pltu
 import matplotlib.patches as patches
 from matplotlib.lines import Line2D
 
-NUKE_COLOR = "#e5cd00"
-THERMAL_COLOR = "#7e52a0"
-WIND_COLOR = "#71cdb8"
-SOLAR_COLOR = "#d66b0d"
-HYDRO_COLOR = "#1f73b5"
-NUKE_ID = 0
-THERMAL_ID = 1
-WIND_ID = 2
-SOLAR_ID = 3
-HYDRO_ID = 4
-TYPE_GEN = {'nuclear': NUKE_ID, 'thermal': THERMAL_ID, 'wind': WIND_ID, 'solar': SOLAR_ID, 'hydro': HYDRO_ID}
-COLOR_GEN = {NUKE_ID: NUKE_COLOR, THERMAL_ID: THERMAL_COLOR, WIND_ID: WIND_COLOR,
-             SOLAR_ID: SOLAR_COLOR, HYDRO_ID: HYDRO_COLOR}
+from grid2op.PlotGrid.config import *  # all colors
 
 
 class GenDraw(patches.CirclePolygon):